+++ src/main/java/org/openlmis/referencedata/repository/FacilityTypeRepository.java
@@ -22,10 +22,12 @@ import org.openlmis.referencedata.domain.FacilityType;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
 @JaversSpringDataAuditable
-public interface FacilityTypeRepository extends JpaRepository<FacilityType, UUID> {
+public interface FacilityTypeRepository extends JpaRepository<FacilityType, UUID>,
+    BaseAuditableRepository<FacilityType, UUID> {
 
   @Override
   <S extends FacilityType> S save(S entity);
@@ -39,4 +41,22 @@ public interface FacilityTypeRepository extends JpaRepository<FacilityType, UUID
   Page<FacilityType> findByActive(Boolean active, Pageable pageable);
 
   boolean existsByCode(@Param("code") String code);
+
+  @Query(value = "SELECT\n"
+      + "    ft.*\n"
+      + "FROM\n"
+      + "    referencedata.facility_types ft\n"
+      + "WHERE\n"
+      + "    id NOT IN (\n"
+      + "        SELECT\n"
+      + "            id\n"
+      + "        FROM\n"
+      + "            referencedata.facility_types ft\n"
+      + "            INNER JOIN referencedata.jv_global_id g "
+      + "ON CAST(ft.id AS varchar) = SUBSTRING(g.local_id, 2, 36)\n"
+      + "            INNER JOIN referencedata.jv_snapshot s  ON g.global_id_pk = s.global_id_fk\n"
+      + "    )\n"
+      + " ORDER BY ?#{#pageable}",
+      nativeQuery = true)
+  Page<FacilityType> findAllWithoutSnapshots(Pageable pageable);
 }