+++ src/main/java/org/openlmis/referencedata/repository/ProcessingScheduleRepository.java
@@ -20,10 +20,32 @@ import java.util.UUID;
 import org.javers.spring.annotation.JaversSpringDataAuditable;
 import org.openlmis.referencedata.domain.Code;
 import org.openlmis.referencedata.domain.ProcessingSchedule;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
 
 @JaversSpringDataAuditable
 public interface ProcessingScheduleRepository
-      extends JpaRepository<ProcessingSchedule, UUID> {
+      extends JpaRepository<ProcessingSchedule, UUID>,
+      BaseAuditableRepository<ProcessingSchedule, UUID> {
   Optional<ProcessingSchedule> findOneByCode(Code code);
+
+  @Query(value = "SELECT\n"
+      + "    ps.*\n"
+      + "FROM\n"
+      + "    referencedata.processing_schedules ps\n"
+      + "WHERE\n"
+      + "    id NOT IN (\n"
+      + "        SELECT\n"
+      + "            id\n"
+      + "        FROM\n"
+      + "            referencedata.processing_schedules ps\n"
+      + "            INNER JOIN referencedata.jv_global_id g "
+      + "ON CAST(ps.id AS varchar) = SUBSTRING(g.local_id, 2, 36)\n"
+      + "            INNER JOIN referencedata.jv_snapshot s  ON g.global_id_pk = s.global_id_fk\n"
+      + "    )\n"
+      + " ORDER BY ?#{#pageable}",
+      nativeQuery = true)
+  Page<ProcessingSchedule> findAllWithoutSnapshots(Pageable pageable);
 }