+++ src/main/java/org/openlmis/referencedata/repository/RequisitionGroupRepository.java
@@ -19,12 +19,33 @@ import java.util.UUID;
 import org.javers.spring.annotation.JaversSpringDataAuditable;
 import org.openlmis.referencedata.domain.RequisitionGroup;
 import org.openlmis.referencedata.repository.custom.RequisitionGroupRepositoryCustom;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
 
 @JaversSpringDataAuditable
 public interface RequisitionGroupRepository
-    extends JpaRepository<RequisitionGroup, UUID>, RequisitionGroupRepositoryCustom {
+    extends JpaRepository<RequisitionGroup, UUID>, RequisitionGroupRepositoryCustom,
+    BaseAuditableRepository<RequisitionGroup, UUID> {
 
   <S extends RequisitionGroup> S findByCode(String code);
 
+  @Query(value = "SELECT\n"
+      + "    r.*\n"
+      + "FROM\n"
+      + "    referencedata.requisition_groups r\n"
+      + "WHERE\n"
+      + "    id NOT IN (\n"
+      + "        SELECT\n"
+      + "            id\n"
+      + "        FROM\n"
+      + "            referencedata.requisition_groups r\n"
+      + "            INNER JOIN referencedata.jv_global_id g "
+      + "ON CAST(r.id AS varchar) = SUBSTRING(g.local_id, 2, 36)\n"
+      + "            INNER JOIN referencedata.jv_snapshot s  ON g.global_id_pk = s.global_id_fk\n"
+      + "    )\n"
+      + " ORDER BY ?#{#pageable}",
+      nativeQuery = true)
+  Page<RequisitionGroup> findAllWithoutSnapshots(Pageable pageable);
 }