+++ src/main/java/org/openlmis/referencedata/repository/RightRepository.java
@@ -19,10 +19,32 @@ import java.util.UUID;
 import org.javers.spring.annotation.JaversSpringDataAuditable;
 import org.openlmis.referencedata.domain.Right;
 import org.openlmis.referencedata.repository.custom.RightRepositoryCustom;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
 @JaversSpringDataAuditable
 public interface RightRepository extends PagingAndSortingRepository<Right, UUID>,
-        RightRepositoryCustom {
+        RightRepositoryCustom,
+        BaseAuditableRepository<Right, UUID> {
   Right findFirstByName(String name);
+
+  @Query(value = "SELECT\n"
+      + "    r.*\n"
+      + "FROM\n"
+      + "    referencedata.rights r\n"
+      + "WHERE\n"
+      + "    id NOT IN (\n"
+      + "        SELECT\n"
+      + "            id\n"
+      + "        FROM\n"
+      + "            referencedata.rights r\n"
+      + "            INNER JOIN referencedata.jv_global_id g "
+      + "ON CAST(r.id AS varchar) = SUBSTRING(g.local_id, 2, 36)\n"
+      + "            INNER JOIN referencedata.jv_snapshot s  ON g.global_id_pk = s.global_id_fk\n"
+      + "    )\n"
+      + " ORDER BY ?#{#pageable}",
+      nativeQuery = true)
+  Page<Right> findAllWithoutSnapshots(Pageable pageable);
 }