+++ src/main/java/org/openlmis/referencedata/repository/SupplyLineRepository.java
@@ -21,16 +21,36 @@ import org.javers.spring.annotation.JaversSpringDataAuditable;
 import org.openlmis.referencedata.domain.Facility;
 import org.openlmis.referencedata.domain.SupplyLine;
 import org.openlmis.referencedata.repository.custom.SupplyLineRepositoryCustom;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
 @JaversSpringDataAuditable
 public interface SupplyLineRepository extends JpaRepository<SupplyLine, UUID>,
-    SupplyLineRepositoryCustom {
+    SupplyLineRepositoryCustom,
+    BaseAuditableRepository<SupplyLine, UUID> {
 
   @Query(name = "SupplyLine.findSupplyingFacilities")
   List<Facility> findSupplyingFacilities(@Param("programId") UUID programId,
                                          @Param("supervisoryNodeId") UUID supervisoryNodeId);
 
+  @Query(value = "SELECT\n"
+      + "    sln.*\n"
+      + "FROM\n"
+      + "    referencedata.supply_lines sln\n"
+      + "WHERE\n"
+      + "    id NOT IN (\n"
+      + "        SELECT\n"
+      + "            id\n"
+      + "        FROM\n"
+      + "            referencedata.supply_lines sln\n"
+      + "            INNER JOIN referencedata.jv_global_id g "
+      + "ON CAST(sln.id AS varchar) = SUBSTRING(g.local_id, 2, 36)\n"
+      + "            INNER JOIN referencedata.jv_snapshot s  ON g.global_id_pk = s.global_id_fk\n"
+      + "    )\n"
+      + " ORDER BY ?#{#pageable}",
+      nativeQuery = true)
+  Page<SupplyLine> findAllWithoutSnapshots(Pageable pageable);
 }